+++ app/src/main/java/woowacourse/shopping/view/core/handler/CartQuantityHandler.kt
@@ -1,0 +1,7 @@
+package woowacourse.shopping.view.core.handler
+
+interface CartQuantityHandler {
+    fun onClickIncrease(productId: Long)
+
+    fun onClickDecrease(productId: Long)
+}
