+++ app/src/main/java/woowacourse/shopping/view/main/adapter/ProductAdapterEventHandler.kt
@@ -1,0 +1,5 @@
+package woowacourse.shopping.view.main.adapter
+
+import woowacourse.shopping.view.core.handler.CartQuantityHandler
+
+interface ProductAdapterEventHandler : ProductAdapter.Handler, CartQuantityHandler
